Complete the project

+++ C#Projects/ExaminationSystem/ExaminationSystem/Classes/Questions.cs
@@ -12,14 +12,14 @@ namespace ExaminationSystem.Classes
         #region Properities
         public string? Header { get; set; }        // Title of the Questions
         public string? Body { get; set; }          // Description of the Questions
-        public decimal Marks { get; set; }         // Question's Marks
+        public int Marks { get; set; }         // Question's Marks
         public Answers[]? AnswerList { get; set; } // all answers (Options for each answer)
         public int RightAnswers { get; set; }      // the right answer of the question
         #endregion
 
         // Will be called each time we create an object from this class
         #region Constructors  
-        public Questions(string? header, string? body, decimal marks, int NumberOfAnswers)
+        public Questions(string? header, string? body, int marks, int NumberOfAnswers)
         {
             this.Header = header;
             this.Body = body;
